Added support for Mill 0.11 (0.11.0-M4 milestone)
Also bumped Scala version to 2.13.10 and scoverage to 2.0.8.
