+++ main/src/de/tobiasroeser/mill/vaadin/VaadinResult.scala
@@ -1,0 +1,11 @@
+package de.tobiasroeser.mill.vaadin
+
+import mill.api.PathRef
+import upickle.default.macroRW
+import upickle.default.ReadWriter
+
+object VaadinResult {
+  implicit val upickleRW: ReadWriter[VaadinResult] = macroRW[VaadinResult]
+}
+
+case class VaadinResult(classesDir: PathRef, productionMode: Boolean)
